+++ lib/atmospheric/export/iso_25331985/table_three_attrs.rb
@@ -2,7 +2,7 @@ module Atmospheric
   module Export
     module Iso25331985
       class TableThreeAttrs < Lutaml::Model::Serializable
-        attribute :pressure_mmhg, :integer
+        attribute :pressure_mmhg, :float
         attribute :geopotential_altitude, :integer
 
         key_value do
